[ADDITION] Interacting with Host elements from Inside Components
As we can see, in the host element we had to write class="control" in the new-ticket component which is the host component of <app-control>.

Problem: For every <app-control> component we have to write "class="control"". It really easy to forget.

Solution: We can remove the "class="control"" and put that attribute inside the host element "@Component" decorator as "host: {
    class: 'control'
  }"

As a result, app-control will render as <app-control class="control"> in the browser.