[ADDITION] Making Sense of Signal Effects
The signal creates a subscription between the variables in your template to the component file.

Problem: What if we want a subcription to the signal value inside the component file?

Answer: effect. Angular automatically tracks which signals are read inside an effect and re-runs the effect when any of those signals changes.